+++ player.h
@@ -21,30 +21,30 @@ private:
     int _constitution;
     int _luck;
 
-    const uint8_t _starting_agility;
-    const uint8_t _starting_constitution;
-    uint8_t _starting_luck;
+    const int _starting_agility;
+    const int _starting_constitution;
+    int _starting_luck;
 
-    uint16_t _gold;
-    uint8_t _rations;
-    uint8_t _elixir_count;
+    int _gold;
+    int _rations;
+    int _elixir_count;
     ElixirType _elixir_type;
 
     std::vector<std::string> _inventory;
 
 public:
-    Player(QObject* parent, uint8_t agility, uint8_t constitution, uint8_t luck, ElixirType elixir_type);
+    Player(QObject* parent, int agility, int constitution, int luck, ElixirType elixir_type);
 
     Q_INVOKABLE int getAgility() const;
     Q_INVOKABLE int getConstitution() const;
     Q_INVOKABLE int getLuck() const;
-    uint16_t getGold() const;
+    int getGold() const;
     bool hasItem(const std::string& item);
 
     void modifyAgility(int value);
     void modifyConstitution(int value);
     void modifyLuck(int value);
-    void modifyGold(int8_t value);
+    void modifyGold(int value);
 
     bool eatRation();
     bool drinkElixir();
